new commit

+++ 1 MAR/THE SECOND PRACTICAL EXERCISE/the second practical exercise.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main () {
-    int seg, horas, minutos, segundos;
+    long int seg, horas, minutos, segundos;
 
 printf("type the number of seconds: ");
 scanf("%d", &seg);
